requirements.txt
pandas
flask
python-dotenv
mysql-connector-python